// src/app/models/case.model.ts
export class ChartOfAccount {
  id: number = 0;
  code: string = "";
  name: string = "";
  openingDebit: number = null;
  openingCredit: number = null;
  arisingDebit: number = null;
  arisingCredit: number = null;
  closingDebit: number = null;
  closingCredit: number = null;
  currency: string = "";
  isForeignCurrency: boolean = false;
  exchangeRate: number = null;
  arisingForeignDebit: number = null;
  arisingForeignCredit: number = null;
  openingForeignDebit: number = null;
  openingForeignCredit: number = null;
  isSpendAccount: boolean = false;
  type: number = 0;
  duration: string = "";
  isProtected: boolean = false;
  displayInsert: boolean = false;
  displayUpdate: boolean = false;
  displayDelete: boolean = false;
  description: string = "";
  accGroup: number = 0;
  classification: number = 0;
  parentRef: string = "";
  isDelete: boolean = false;
  hasChild: boolean = false;
  hasDetails = false;
  minimumStockQuantity = null;
  maximumStockQuantity = null;
  stockCostPrice = null;
  stockSellingPrice = null;
  stockUnitPrice = null;
  stockUnit = null;
  warehouseCode = null;
  openingStockQuantity = null;
  closingStockQuantity = null;
  details: ChartOfAccount[] = [];
  openingForeignDebitNB: number = null;
  openingForeignCreditNB: number = null;
  openingStockQuantityNB = null;
  stockUnitPriceNB = null;
  openingDebitNB: number = null;
  openingCreditNB: number = null;
  isInternal: boolean = false;

  constructor(account?: ChartOfAccount) {
    Object.assign(this, account);
  }
}